flaskapp/requirements.txt
tensorflow==2.5.0
keras
Pillow
numpy